apps/backend/requirements.txt
# Phase 1 — Core backend
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
httpx>=0.26.0
pytest>=7.4.0
readability-lxml>=0.8.1
markdownify>=0.11.6
beautifulsoup4>=4.12.0
lxml>=5.0.0
gspread>=6.0.0
google-auth>=2.0.0
psutil>=5.9.0
respx>=0.21.0
llama-cpp-python>=0.2.0
